Update version and add new features
Updated the version in Cargo.toml from 0.1.0 to 0.2.0. Added two new commands to the bot in the src/main.rs file, namely 'AddLink' and 'CheckSite'. Implemented the functions for the commands to add a link to the database and to check a site, respectively. This allows users to add a link to the bot for hourly checks and to request the bot to check the accessibility of a site.