Add timeout for website requests and handle errors
Set a timeout for website requests in both get_request_code and get_site_information functions. If a website fails to respond within 15 seconds, the request will timeout. The application has also been updated to handle timeouts and other request errors. If a request fails, a corresponding error message is displayed to the user. Underlying version is updated to v0.3.0 as well.